fixed syntax and logic errors
fixed syntax errors such as numberofdigits not defined but was defined as numdigits. Also the program wanted to return false, but it is return 0.  The logic errors were that the modulus initiall began at 0 but you cannot divide by 0 so we began at 1. Also, the numDigits was not large enough to compare that i<= numDigits the loop will stop. So I increased that by adding 3 to the numDigits.